Fixed bug when creating a player

+++ PDC Project/src/pdc/project/Model/Archer.java
@@ -5,8 +5,8 @@ import java.util.HashMap;
 
 
 public class Archer extends Player {
-        public Archer(int damage,String name, int health, int exp, ArrayList inventory,
-            int score, HashMap equipped, double hitChance, double critChance, 
+        public Archer(int damage, String name, int health, int exp, ArrayList inventory,
+            double score, HashMap equipped, double hitChance, double critChance, 
             double dodgeChance, boolean defensive, boolean dot) {
         super(damage,name,health,exp,inventory,score,equipped,hitChance,critChance,dodgeChance, defensive, dot);
     }
