+++ PDC Project/src/pdc/project/Model/Warrior.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 
 public class Warrior extends Player {
     public Warrior(int damage, String name, int health, int exp, ArrayList inventory,
-            int score, HashMap equipped, double hitChance, double critChance, 
+            double score, HashMap equipped, double hitChance, double critChance, 
             double dodgeChance, boolean defensive, boolean dot) {
             
         super(damage,name,health,exp,inventory,score,equipped,hitChance,critChance,dodgeChance,defensive,dot);
